Move filters to config file
