+++ os_site/.eleventy.js
@@ -1,7 +1,4 @@
 const eleventyNavigationPlugin = require("@11ty/eleventy-navigation");
-const minFilter = require ('./src/filters/min')
-const headFilter = require ('./src/filters/head')
-const excerptFilter = require('./src/filters/excerpt')
 const { DateTime } = require('luxon');
 
 module.exports = function(eleventyConfig) {
@@ -14,6 +11,26 @@ module.exports = function(eleventyConfig) {
       return DateTime.fromJSDate(date, { zone: 'utc' }).toFormat(
           'dd LLL yyyy'
       );
+    });
+
+    eleventyConfig.addFilter('excerpt', (post) => {
+        const content = post.replace(/(<([^>]+)>)/gi, '');
+        return content.substr(0, content.lastIndexOf(' ', 200)) + '...';
+    });
+
+    eleventyConfig.addFilter('head', (array, n) => {
+        if( n < 0 ) {
+          return array.slice(n);
+        }
+        return array.slice(0, n);
+    });
+
+    eleventyConfig.addFilter('min', (num, minimum) => {
+        if (num < minimum) {
+            return num;
+        } else {
+            return minimum;
+        }
     });
 
     eleventyConfig.setTemplateFormats([
