--- os_site/src/filters/excerpt.js
@@ -1,6 +1,0 @@
-function excerptFilter(post) {
-    const content = post.replace(/(<([^>]+)>)/gi, '');
-    return content.substr(0, content.lastIndexOf(' ', 200)) + '...';
-}
-
-module.exports = excerptFilter
