--- os_site/src/filters/min.js
@@ -1,9 +1,0 @@
-function minFilter(num, minimum) {
-    if (num < minimum) {
-        return num;
-    } else {
-        return minimum;
-    }
-}
-
-module.exports = minFilter
